# openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "ExactBag - API de Integração para Parceiros",
    "description": "## Como funciona\n\nA ExactBag oferece um serviço digital de identificação e proteção de bagagem para passageiros.\n\n### Fluxo da integração:\n1. **Parceiro envia a venda** → `POST /webhooks/sales` com dados do passageiro\n2. **ExactBag processa** → Gera formulário personalizado e envia ao passageiro (email + WhatsApp)\n3. **Passageiro preenche** → Registra fotos e dados da bagagem no formulário\n4. **Parceiro é notificado** → Recebe callback no webhook cadastrado quando o registro é concluído\n5. **Cancelamento (opcional)** → `POST /webhooks/sales/{saleId}/cancel` calcula estorno automaticamente\n\n### Autenticação\nTodas as chamadas (exceto `/health`) exigem header:\n```\nAuthorization: Bearer SUA_API_KEY\n```\nA API Key é fornecida pela ExactBag no momento do cadastro do parceiro.\n\n### Suporte\ncontato@exactbag.com.br",
    "version": "2.0.0",
    "contact": {
      "name": "ExactBag",
      "email": "contato@exactbag.com.br",
      "url": "https://exactbag.com.br"
    }
  },
  "servers": [
    {
      "url": "https://app.exactbag.com.br",
      "description": "Produção"
    }
  ],
  "security": [
    {
      "BearerAuth": []
    }
  ],
...
  "paths": {
    "/webhooks/sales": {
      "post": {
        "tags": ["Vendas"],
        "summary": "Enviar uma nova venda",
        "description": "Envia os dados de uma venda de bagagem. A ExactBag irá:\n- Gerar um formulário personalizado para o passageiro\n- Enviar link do formulário por email e WhatsApp\n- Notificar o parceiro quando o passageiro completar o registro\n\n**Processamento assíncrono**: a resposta retorna um `jobId` para acompanhamento.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaleRequest"
              },
              "example": {
                "partnerId": "INFOTRAVEL",
                "saleId": "RES-2026-78901",
                "customerName": "Maria Santos",
                "customerEmail": "maria.santos@email.com",
                "customerPhone": "+5511988887777",
                "productCode": "5383660",
                "roundTrip": true,
                "outboundDate": "2026-04-15",
                "returnDate": "2026-04-22",
                "baggageQty": 2,
                "hasInsurance": true
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "Venda aceita para processamento",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": {
                      "type": "string",
                      "example": "Venda recebida e em processamento."
                    },
                    "jobId": {
                      "type": "string",
                      "format": "uuid",
                      "example": "11111111-1111-4111-8111-111111111111",
                      "description": "Use este ID para consultar o status em GET /jobs/{jobId}"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "error": {
                      "type": "string",
                      "example": "Validation failed"
                    },
                    "details": {
                      "type": "array",
                      "items": { "type": "string" },
                      "example": [
                        "customerEmail is required",
                        "outboundDate must be a valid date"
                      ]
                    }
                  }
                }
              }
            }
          },
          "401": { "description": "API Key ausente ou inválida" },
          "429": {
            "description": "Limite de requisições excedido (máx. 30/minuto)"
          }
        }
      }
    },
...
    "/webhooks/register": {
      "post": {
        "tags": ["Webhook do Parceiro"],
        "summary": "Cadastrar URL de callback",
        "description": "Define a URL que receberá notificações quando eventos ocorrerem.\n\n**Eventos enviados à sua URL:**\n\n1. **`customer.registration.completed`** — Passageiro completou o registro de bagagem:\n```json\n{\n  \"event\": \"customer.registration.completed\",\n  \"saleId\": \"clxyz123abc\",\n  \"partnerId\": \"INFOTRAVEL\",\n  \"customerName\": \"Maria Santos\",\n  \"customerEmail\": \"maria.santos@email.com\",\n  \"submissionId\": \"clxyz456def\",\n  \"completedAt\": \"2026-04-10T14:30:00Z\",\n  \"signature\": \"hmac-sha256-hex...\"\n}\n```\n\n2. **`sale.cancelled`** — Venda foi cancelada:\n```json\n{\n  \"event\": \"sale.cancelled\",\n  \"saleId\": \"RES-2026-78901\",\n  \"partnerId\": \"INFOTRAVEL\",\n  \"customerName\": \"Maria Santos\",\n  \"customerEmail\": \"maria.santos@email.com\",\n  \"cancelledAt\": \"2026-04-08T10:00:00Z\",\n  \"refundType\": \"integral\",\n  \"reason\": \"Cliente desistiu\",\n  \"signature\": \"hmac-sha256-hex...\"\n}\n```\n\n**Verificação da assinatura (recomendado):**\nO campo `signature` é gerado com HMAC-SHA256 usando sua API Key como secret sobre o payload JSON. Também enviado no header `X-ExactBag-Signature`.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["webhookUrl"],
                "properties": {
                  "webhookUrl": {
                    "type": "string",
                    "format": "uri",
                    "example": "https://api.seusite.com.br/callbacks/exactbag"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Webhook registrado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string", "example": "success" },
                    "message": {
                      "type": "string",
                      "example": "Webhook registrado com sucesso"
                    },
                    "webhookUrl": {
                      "type": "string",
                      "example": "https://api.seusite.com.br/callbacks/exactbag"
                    }
                  }
                }
              }
            }
          },
          "400": { "description": "URL inválida" },
          "401": { "description": "API Key ausente ou inválida" }
        }
      }
    },
...
    "/webhooks/unregister": {
      "delete": {
        "tags": ["Webhook do Parceiro"],
        "summary": "Remover webhook",
        "description": "Remove a URL de callback. Sem webhook, o parceiro não receberá notificações automáticas.",
        "responses": {
          "200": { "description": "Webhook removido com sucesso" },
          "401": { "description": "API Key ausente ou inválida" }
        }
      }
    },
    "/webhooks/sales/{saleId}/cancel": {
      "post": {
        "tags": ["Cancelamento"],
        "summary": "Cancelar uma venda",
        "description": "Cancela uma venda e calcula o estorno automaticamente.\n\n**Regras de estorno:**\n- **≥48h antes do voo** → Estorno integral (100%)\n- **<48h antes do voo** → Estorno parcial (50%)\n- **Status `reserva` (não paga)** → Sem estorno\n\n**O que acontece ao cancelar:**\n1. Status da venda muda para `cancelada`\n2. Link do formulário é invalidado (cliente não consegue mais acessar)\n3. Notificações de véspera são canceladas\n4. Callback `sale.cancelled` é enviado ao parceiro (se webhook configurado)",
        "parameters": [
          {
            "name": "saleId",
            "in": "path",
            "required": true,
            "description": "ID da venda no sistema do parceiro (mesmo `saleId` enviado no POST /webhooks/sales)",
            "schema": { "type": "string", "example": "RES-2026-78901" }
          }
        ],
        "requestBody": {
          "required": false,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "reason": {
                    "type": "string",
                    "description": "Motivo do cancelamento (opcional)",
                    "example": "Cliente desistiu da viagem"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Venda cancelada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": { "type": "boolean", "example": true },
                    "message": {
                      "type": "string",
                      "example": "Venda cancelada com sucesso"
                    },
                    "sale": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "string" },
                        "saleId": {
                          "type": "string",
                          "example": "RES-2026-78901"
                        },
                        "customerName": {
                          "type": "string",
                          "example": "Maria Santos"
                        },
                        "status": { "type": "string", "example": "cancelada" },
                        "cancelledAt": {
                          "type": "string",
                          "format": "date-time"
                        }
                      }
                    },
                    "refundType": {
                      "type": "string",
                      "enum": ["integral", "parcial", "nenhum"],
                      "example": "integral"
                    }
                  }
                }
              }
            }
          },
          "404": { "description": "Venda não encontrada" },
          "409": { "description": "Venda já está cancelada" },
          "422": { "description": "Status da venda não permite cancelamento" },
          "401": { "description": "API Key ausente ou inválida" }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "SaleRequest": {
        "type": "object",
        "required": [
          "partnerId",
          "saleId",
          "customerName",
          "customerEmail",
          "customerPhone"
        ],
        "properties": {
          "partnerId": {
            "type": "string",
            "description": "Identificador do parceiro (fornecido pela ExactBag)",
            "example": "INFOTRAVEL"
          },
          "saleId": {
            "type": "string",
            "description": "ID da venda no sistema do parceiro (usado para evitar duplicidade)",
            "example": "RES-2026-78901"
          },
          "customerName": {
            "type": "string",
            "description": "Nome completo do passageiro",
            "example": "Maria Santos"
          },
          "customerEmail": {
            "type": "string",
            "format": "email",
            "description": "Email do passageiro (receberá o link do formulário)",
            "example": "maria.santos@email.com"
          },
          "customerPhone": {
            "type": "string",
            "description": "Telefone com DDI+DDD (receberá WhatsApp)",
            "example": "+5511988887777"
          },
          "roundTrip": {
            "type": "boolean",
            "description": "Viagem de ida e volta?",
            "default": false,
            "example": true
          },
          "outboundDate": {
            "type": "string",
            "format": "date",
            "description": "Data de ida (formato YYYY-MM-DD)",
            "example": "2026-04-15"
          },
          "returnDate": {
            "type": "string",
            "format": "date",
            "description": "Data de volta (apenas se roundTrip=true)",
            "example": "2026-04-22"
          },
          "baggageQty": {
            "type": "integer",
            "description": "Quantidade de bagagens contratadas",
            "default": 1,
            "minimum": 1,
            "example": 2
          },
          "hasInsurance": {
            "type": "boolean",
            "description": "Indica se a venda inclui seguro de bagagem. Propaga para o registro do passageiro e fica visível no painel.",
            "default": false,
            "example": true
          },
          "productCode": {
            "type": "string",
            "description": "Product code for the sale. To trigger the physical TAG flow, provide `5383660` (Essencial X2) or `5383661` (Cover X2). If omitted, the sale uses the digital flow.",
            "enum": ["5383660", "5383661"],
            "example": "5383660"
          }
        }
      },
...
    },
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Informe sua API Key: Bearer SUA_API_KEY"
      }
    }
  }
}